update hardwareclassification controller
Use a finalizer to control when it is safe to delete the profile. We
want to wait for all hosts using the label to be updated so we don't
leave extra labels on a host.

Update the status to show whether the profile has matched anything or
not.

Reconcile the hardware classification resource when a host changes so
we can update the status field.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>